requirements.txt

fastapi
uvicorn
pydantic>=2
openai>=1.30.0
python-dotenv
